hack: testing -Wl,-fuse-ld=gold does not work correctly on hardened
Not sure why this test falsely passes with the hardened compiler when it
normally will report the following error:

    ld: -f may not be used without -shared

But apparently the default options hardened uses makes interpreting the
option as -f valid usage. For reference the option is:

    -f name
    --auxiliary=name
        When creating an ELF shared object, set the internal DT_AUXILIARY
        field to the specified name.  This tells the dynamic linker that
        the symbol table of the shared object should be used as an
        auxiliary filter on the symbol table of the shared object name.

This in turn causes a stray library to show up in ldd output:

    use-ld=gold => not found

Which seems mostly harmless but does cause some confusion.