networkd: do not drop foreign configs
this reverts a change in networkd behavior that drops foreign configs
(such as IPs and routes) of interfaces that have networkd configs.

this causes issues with interfaces created and configured by
non-networkd applications, such as the bridge that docker creates and
configured.

the intention is to keep compatibility with previous networkd behavior
until a better solution is found.