nspawn: decouple --boot from CLONE_NEWIPC
This commit is a minor tweak after the split of `--share-system`,
decoupling the `--boot` option from IPC namespacing.

Historically there has been a single `--share-system` option for
sharing IPC/PID/UTS with the host, which was incompatible with
boot/pid1 mode. After the split, it is now possible to express the
requirements with better granularity.